ExtJS6ByExample/ext6project/requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.3.1
django-coverage-plugin==1.3.1
Sphinx==1.5.1
django-extensions==1.7.5
Werkzeug==0.11.15
django-test-plus==1.0.16
factory-boy==2.8.1

django-debug-toolbar==1.6

# improved REPL
ipdb==0.10.1

pytest-django==3.1.2
pytest-sugar==0.8.0

fabric==1.13.1
